// komg/komgo-member-client/src/features/receivable-discounting-legacy/resources/strings.ts
export enum Strings {
  ApplyForDiscountingRiskCoverActionText = 'Apply for risk cover / discounting',
  ReceivableDiscountingData = 'Receivable discounting data',
  RiskCoverReceivableDiscountingData = 'Risk cover / receivable discounting data',
  RiskCoverData = 'Risk cover data',

  // HEADER Specific titles
  ReceivableDiscountingAndRiskCoverSimpleDisplayTitle = 'Risk cover / receivable discounting',
  RiskCoverSimpleDisplayTitle = 'Risk cover',
  ReceivableDiscountingSimpleDisplayTitle = 'Receivable discounting',

  RiskCoverDescription = 'No discounting option',
  RiskCoverReceivableDiscountingDescription = 'Cover the risk with the option to request discounting',
  ReceivableDiscountingDescription = 'Discount with or without recourse',

  DiscountWithoutRecourseTitle = 'Discount without recourse',
  DiscountBlendedTitle = 'Discount blended',
  DiscountWithRecourseTitle = 'Discount with recourse',
  DiscountTitle = 'Discount',
  RiskCoverOnlyTitle = 'Risk cover only',
  RiskCoverWithDiscountingOptionTitle = 'Risk cover with discount option',

  TitleTransferredDateTitle = 'Date',
  TitleToBeTransferredDateTitle = 'Estimated date',
  IndicativeCostOfFundsTitle = 'COF',
  IndicativeCostOfFundsMediumTitle = 'Indicative COF',
  AllInPricing = 'All in pricing',
  AddOnOverLiborTitle = 'Add on over Libor',

  StraightDiscount = 'Straight discount',
  DiscountToYield = 'Discount to yield',
  OtherFormula = 'Other formula',

  Maturity1Day = 'Overnight',
  Maturity7Days = '1 week',
  Maturity30Days = '1 month',

  SellerTradeStatusDefault = 'To be covered/discounted',

  RequestTabName = 'Request',
  CreditAppetiteTabName = 'Credit appetite'
}